% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.R
\name{as.data.frame.ejObject}
\alias{as.data.frame.ejObject}
\title{convert an ejObject to a dataframe with one row per record}
\usage{
\method{as.data.frame}{ejObject}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{an ejObject}

\item{row.names}{NULL or a character vector giving the row names for the data frame. Missing values are not allowed.}

\item{optional}{not used.}

\item{...}{other parameters passed to \code{\link{data.frame}}.}
}
\value{
dataframe
}
\description{
convert an ejObject to a dataframe with one row per record
}

