% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repijson.R
\name{as.ejObject}
\alias{as.ejObject}
\title{generic as function}
\usage{
as.ejObject(x, ...)
}
\arguments{
\item{x}{an object to convert to repijson}

\item{...}{other parameters to pass to the converter}
}
\description{
generic as function
}

