% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSONGenerator.R
\name{attributeAsJSON}
\alias{attributeAsJSON}
\title{Convert an attribute into JSON}
\usage{
attributeAsJSON(attribute)
}
\arguments{
\item{attribute}{The attribute object to be converted to JSON}
}
\description{
Convert an attribute into JSON
}

