% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fileOps.R
\name{attributeParser}
\alias{attributeParser}
\title{convert a list representing an attribute into an ejAttribute}
\usage{
attributeParser(x)
}
\arguments{
\item{x}{A list with name, type and value components may optionally include
a unit component}
}
\value{
an ejAttribute
}
\description{
convert a list representing an attribute into an ejAttribute
}

