% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binaryAttributes.R
\name{attributeToFile}
\alias{attributeToFile}
\title{Convert a base64 attribute to a file}
\usage{
attributeToFile(attribute, file)
}
\arguments{
\item{attribute}{The ejAttribute to convert}

\item{file}{The filename to write to}
}
\value{
invisilbe NULL
}
\description{
Convert a base64 encoded attribute to a file.
}
\author{
Thomas Finnie (thomas.finnie@phe.gov.uk)
}

