% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fileOps.R
\name{geojsonListToSp}
\alias{geojsonListToSp}
\title{Convert a json type list into a spatial object}
\usage{
geojsonListToSp(x)
}
\arguments{
\item{x}{a list representing a geojson object}
}
\description{
Convert a json type list into a spatial object
}

