% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{create_ejMetadata}
\alias{create_ejMetadata}
\title{Create metadata}
\usage{
create_ejMetadata(attributes)
}
\arguments{
\item{attributes}{list of attributes of the metadata}
}
\value{
an ejMetadata object
}
\description{
This function defines epiJSON Metadata
output \code{ejMetadata}
}

