% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/constructors.R
\name{create_ejObject}
\alias{create_ejObject}
\title{Create an object}
\usage{
create_ejObject(metadata, records)
}
\arguments{
\item{metadata}{metadata for the whole ejObject}

\item{records}{list of records}
}
\value{
an ejObject object
}
\description{
This function defines epiJSON objects
output \code{ejObject}
}

