% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{createStandardMetadata}
\alias{createStandardMetadata}
\title{Generate standard metadata for EpiJSON}
\usage{
createStandardMetadata(title, generatorName,
  generatorLongName = generatorName, generatorVersion = "", dataSource,
  institution = "", history = "", runUUID = generateUUID(),
  parameters = list(), references = "", comment = "")
}
\arguments{
\item{title}{A succinct description of what is in the file}

\item{generatorName}{The short name of the program. This should be URL friendly.
This will appear in the 'generatorName' attribute and may be used
in other systems to provide web access to results.}

\item{generatorLongName}{Ideally this is descriptive of what the generator is
and does. This will appear in the 'history' attribute.}

\item{generatorVersion}{The version of the generator.}

\item{dataSource}{Free text for the source parameter. Where did this data come from?}

\item{institution}{Where this data was produced?}

\item{history}{This attribute within the EpiJSON file provides an audit trail
for the file. Provide previous history here, a timestamp followed by the generator name,
generator version, username and machine name will be added at the bottom.}

\item{runUUID}{a Universally Unique IDentifier for this generator run. Must follow
the Open Software Foundation UUID standard. Will be used for the runUUID
attribute.}

\item{parameters}{The parameters that the generator was run with. A named list
is expected and this will be used to fill in the parameters attributes.}

\item{references}{Published or web based references that describe the data or
the methods used to produce it.}

\item{comment}{Free text for additional comments.}
}
\description{
Provide an ejMetadata object with standard metadata. In large part these
follow the reccomendations of the NetCDF CF convention with a few additional
parameters to help with dataset and model validation.
}
\note{
This function will generate the following attributes:
\itemize{
 \item{title} {A brief title for the dataset.}
 \item{institution} {Where the dataset was produced.}
 \item{source} {The source of the data the generator used.}
 \item{history} {The history of the dataset.}
 \item{references} {Published references for data or methods.}
 \item{comment} {Free text commentary on the data.}
 \item{runUUID} {A unique identifier for a run.}
 \item{parameters} {Attributes containing parameters used for
 this generator run (they take their name from the names in the list with
 paramer_ prepended).}
 \item{runtime} {A record of the R session (version and attached libraries)
  at the point of invoking this command.}
}
}
\author{
Tom Finnie (Thomas.Finnie@phe.gov.uk)
}

