% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.frame.R
\name{define_ejEvent}
\alias{define_ejEvent}
\title{Creates a event definition}
\usage{
define_ejEvent(id = NA, name = NULL, date = NULL, location = NULL,
  attributes = NULL)
}
\arguments{
\item{id}{A character string naming the column that defines the id for a
event. May be NA, and if so will be automatically generated.}

\item{name}{Either a character string with the event name. Or the name of the
column that contains names for events as a character string. If the column
name is found in the input data.frame to \code{\link{as.ejObject.data.frame}}
then the data in the column is used otherwise the string itself is used.
Event names might be things such as infection, swab, hospital admission,etc.}

\item{date}{A character string naming the column that defines the date
an event occured. This should be in POSIXct format. May be NA.}

\item{location}{A list with entities x, y and proj4string. x and y should be
character strings naming the columns where the x and y of the location are
defined. crs may be "" or a proj4string.}

\item{attributes}{A character vector naming the columns for attributes of the
event. The attributes will be named after the columns, with type taken from
column type.}
}
\description{
Simplifies the definition of events from columns within a dataframe
}

