% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extractors.R
\name{ejAttributes}
\alias{ejAttributes}
\title{Extract the attributes from an EpiJson Object}
\usage{
ejAttributes(x)
}
\arguments{
\item{x}{An EpiJSON object (class ejObject, ejRecord or ejEvent)}
}
\value{
A list of attributes
}
\description{
Get the attributes as a list from an EpiJSON Object
}

