% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extractors.R
\name{ejRecords}
\alias{ejRecords}
\title{Extract the records from an EpiJson Object}
\usage{
ejRecords(x)
}
\arguments{
\item{x}{An EpiJSON object (class ejObject)}
}
\value{
A list of records
}
\description{
Get the records as a list from an EpiJSON Object
}
\note{
This is functionally equivelent to calling x$records (which may be
 quicker as therer is no object type checking) but open to change in later
 versions.
}

