% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileOps.R
\name{eventParser}
\alias{eventParser}
\title{convert a list representing an event into an ejEvent}
\usage{
eventParser(x)
}
\arguments{
\item{x}{A list with id, name, attributes and one or both of date/location 
components.}
}
\value{
an ejRecord
}
\description{
convert a list representing an event into an ejEvent
}

