% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryAttributes.R
\name{fileToAttribute}
\alias{fileToAttribute}
\title{Convert a file to a base64 encoded attribute}
\usage{
fileToAttribute(file, name = as.character(file))
}
\arguments{
\item{file}{The file to read and convert}

\item{name}{The name for the attribute}
}
\value{
An ejAttribute containing the file
}
\description{
Read a file in binary mode and convert the bytesteam into a base64 encoded 
ejAttribute
}
\author{
Thomas Finnie (thomas.finnie@phe.gov.uk)
}

