% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UUID.r, R/helpers.R
\name{generateUUID}
\alias{generateUUID}
\title{Generate a UUID}
\usage{
generateUUID()

generateUUID()
}
\description{
Use R's random number generator to create a version 4 UUID following RFC4122

Use R's random number generator to create a version 4 UUID
}
\examples{
generateUUID()
generateUUID()
}
\author{
Tom Finnie (Thomas.Finnie@phe.gov.uk)

Tom Finnie (Thomas.Finnie@phe.gov.uk)
}

