% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UUID.r
\name{is.UUID}
\alias{is.UUID}
\title{Check strings for conformance to RFC4122}
\usage{
is.UUID(x)
}
\arguments{
\item{x}{a character vector of strings to check for conformance}
}
\description{
Check to see that strings conform to the UUID specification
}
\examples{
is.UUID(generateUUID())

is.UUID(replicate(5, generateUUID()))
}
\author{
Tom Finnie (Thomas.Finnie@phe.gov.uk)
}

