% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{listToAttributes}
\alias{listToAttributes}
\title{Convert a list with named elements into attributes}
\usage{
listToAttributes(x)
}
\arguments{
\item{x}{A list with named elements}
}
\value{
result A list of attributes that comprise the list
}
\description{
Take a list and generate ejAttributes from it
}
\examples{
 test <- list(testNumber=1, 5, aword="meow")
 listToAttributes(test)
}
\author{
Thomas.Finnie (Thomas.Finnie@phe.gov.uk
}

