% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{notNA}
\alias{notNA}
\title{Return a value only if another is not NA}
\usage{
notNA(x, trueValue, defaultValue = NA)
}
\arguments{
\item{x}{The value to test for NA}

\item{trueValue}{The value to return if x is not NA}

\item{defaultValue}{The value to return if x is NA}
}
\value{
defaultValue if x is NA or trueValue if x is not NA
}
\description{
Return a value only if another is not NA
}
\note{
Not terribly different from an ifelse but is more graceful with NULL 
 values
}

