% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{+.ejMetadata}
\alias{+.ejMetadata}
\title{Add an object to a metadata object}
\usage{
\method{+}{ejMetadata}(metadata, x)
}
\arguments{
\item{metadata}{an ejMetadata object}

\item{x}{the object to be added}
}
\description{
Add an object to a metadata object
}

