% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.R
\name{print.ejAttribute}
\alias{print.ejAttribute}
\title{print an ejAttribute object}
\usage{
\method{print}{ejAttribute}(x, ...)
}
\arguments{
\item{x}{An ejAttribute object}

\item{...}{Other arguments to print (not used)}
}
\description{
print an ejAttribute object
}

