% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ejEvent}
\alias{print.ejEvent}
\title{print an ejEvent object}
\usage{
\method{print}{ejEvent}(x, ...)
}
\arguments{
\item{x}{An ejEvent object}

\item{...}{Other arguments to print (not used)}
}
\description{
print an ejEvent object
}

