% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.R
\name{print.ejMetadata}
\alias{print.ejMetadata}
\title{print an ejMetadata object}
\usage{
\method{print}{ejMetadata}(x, ...)
}
\arguments{
\item{x}{An ejMetadata object}

\item{...}{Other arguments to print (not used)}
}
\description{
print an ejMetadata object
}

