% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ejObject}
\alias{print.ejObject}
\title{print an ejObject object}
\usage{
\method{print}{ejObject}(x, ...)
}
\arguments{
\item{x}{An ejObject object}

\item{...}{Other arguments to print (not used)}
}
\description{
print an ejObject object
}

