% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fileOps.R
\name{read.epijson}
\alias{read.epijson}
\title{convert a textfile in epiJSON form into an ejObject}
\usage{
read.epijson(file)
}
\arguments{
\item{file}{an epiJSON filename or string to convert to R}
}
\value{
An ejObject
}
\description{
Takes an epiJSON string in a connection (file) and converts to an ejObject.
}
\examples{
listJSON <- read.epijson( system.file("extdata//example.JSON", package="repijson"))
}

