% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fileOps.R
\name{recordParser}
\alias{recordParser}
\title{convert a list representing a record into an ejRecord}
\usage{
recordParser(x)
}
\arguments{
\item{x}{A list with id, attributes and events components}
}
\value{
an ejRecord
}
\description{
convert a list representing a record into an ejRecord
}

