% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repijson.R
\docType{package}
\name{repijson}
\alias{repijson}
\alias{repijson-package}
\title{Classes and functions to handle EpiJSON files}
\description{
EpiJSON is a universal JSON format for epidemiological data. More information on this format can be found at:
http://github.com/Hackout2/EpiJSON
}
\details{
repijson is a package implementing classes and functions for importing, exporting and converting EpiJSON files.
}

