% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fileOps.R
\name{write.epijson}
\alias{write.epijson}
\title{Write out an ejObject}
\usage{
write.epijson(x, pretty = FALSE, file = "", fileEncoding = "",
  append = FALSE)
}
\arguments{
\item{x}{an ejObject to be written}

\item{pretty}{Should the JSON be prittified (made more human readble with
whitespace)?}

\item{file}{either a character string naming a file or a connection open
for writing.  ‘""’ indicates output to the console.}

\item{fileEncoding}{character string: if non-empty declares the encoding to
be used on a file (not a connection) so the character data
can be re-encoded as they are written.  See ‘file’.}

\item{append}{Should the output be appended?}
}
\description{
Write an ejObject out to a file or connection in JSON form
}
\author{
Thomas Finnie (Thomas.Finnie@phe.gov.uk)
}

