% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{create_ejRecord}
\alias{create_ejRecord}
\title{Create a record}
\usage{
create_ejRecord(id, attributes, events)
}
\arguments{
\item{id}{This is the unique identifier of the record (a valid UUID)}

\item{attributes}{list of attributes associated with this record}

\item{events}{list of events associated with this record}
}
\value{
an ejRecord object
}
\description{
This function defines records
output \code{ejRecord}
}
\examples{
#somewhere on South Bank
demoPoints <- sp::SpatialPoints(data.frame(lat=51.4982778, long=-0.0975535), 
	proj4string=sp::CRS("+init=epsg:4326"))
}

