% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{generateUUID}
\alias{generateUUID}
\title{Generate a UUID}
\usage{
generateUUID()
}
\description{
Use R's random number generator to create a version 4 UUID
}
\examples{
generateUUID()
}
\author{
Tom Finnie (Thomas.Finnie@phe.gov.uk)
}

