% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ejRecord}
\alias{print.ejRecord}
\title{print an ejRecord object}
\usage{
\method{print}{ejRecord}(x, ...)
}
\arguments{
\item{x}{An ejRecord object}

\item{...}{Other arguments to print (not used)}
}
\description{
print an ejRecord object
}

