% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repijson.R
\docType{package}
\name{repijson}
\alias{repijson}
\alias{repijson-package}
\title{A data format for epidemiological data}
\description{
a JSON format for epidemiological data with functions to convert to and from other useful formats
}

